Added a more meaningful message to the SD Info screen if the SD card hasn't been mounted. Added compression of VGMs prior to uploading to SPC RAM, so larger VGMs can now be played (up to about 90kB).

+++ trunk/SNES/NeoMythMenu/common.h
@@ -174,7 +174,7 @@ extern void print_cheat_list();
 extern void print_hw_card_rev();
 extern void printxy(char *, u16, u16, u16, u16);
 extern void print_hex(u8, u16, u16, u16);
-extern void print_dec(u16, u16, u16, u16);
+extern void print_dec(DWORD, u16, u16, u16);
 extern void set_printxy_clip_rect(u16 x1, u16 y1, u16 x2, u16 y2);
 
 extern void show_scroll_indicators();
